Move SYS_DPAA_QBMAN to Kconfig
The CONFIG_SYS_DPAA_QBMAN define is used by DPAA1 freescale SOCs to
add device tree fixups that allow deep sleep in Linux. The define was
placed in header files included by a number of boards, but was not
explicitly documented in any of the Kconfigs. A description was added
to the drivers/networking menuconfig and default selection for
current SOCs that have this part

Signed-off-by: Ahmed Mansour <ahmed.mansour@nxp.com>

Series-version: 5

Series-changes: 4
- Add QBMan support in Kconfig auto select for B4420 and T4160 SOCs

Series-changes: 5
- Add net device to T1024QDS_DDR4_defconfig to enable QBMan fixup
- Remove QBMan define in kmp204x-common.h and rely on SOC auto select

Cover-letter:
arm: ppc: Share DPAA1 fixups between ARCHs
- Support DPAA1 QBMan device tree fixups in a shared location for both
  arm and ppc architectures
- cleanup a define from header files and add as Kconfig with auto
  selection based on SOC
END

Series-to: Joe Hershberger <joe.hershberger@ni.com>
Series-to: York Sun <york.sun@nxp.com>
Series-to: Roy Pledge <roy.pledge@nxp.com>
Series-to: Prabhakar Kushwaha <prabhakar.kushwaha@nxp.com>