spl: imx: Select simple malloc in SPL for SPL_OPTEE_BOOT
This patch is part of the i.MX Windows 10 IoT Core boot flow.

If CONFIG_SYS_SPL_MALLOC_START is defined, it will override
the CONFIG_SPL_SYS_MALLOC_SIMPLE and allocate space in DRAM instead of
on-chip ram. This patch prevents the definition of those values for
i.MX6 and i.MX7 SPL if CONFIG_SPL_OPTEE_BOOT is selected.

Signed-off-by: Henry Beberman <henry.beberman@microsoft.com>