mx6cuboxi: Add Windows boot support for mx6cuboxi
This patch adds a new bootable configuration for Windows 10 IoT Core on
the i.MX6 Quad HummingBoard.

It also adds a new device tree sources file for imx6qdl-mx6cuboxi.dts.
This is not required by Windows or UEFI but the build system requires a
device tree when building SPL with flat image tree support.

Signed-off-by: Henry Beberman <henry.beberman@microsoft.com>