imx: Add get_random_bytes function to FSL CAAM driver
The FSL CAAM driver currently initializes the RNG hardware, but does not
provide a function to retrieve random data from it. This patch adds in a
function to fill a buffer with random data from the CAAM. It also adds
the config SPL_FSL_CAAM to signal that the CAAM should be initialized
for use in SPL.

In a future patch we plan to make use of hardware generated random
numbers in SPL to generate a device identity certificate chain.

Signed-off-by: Henry Beberman <henry.beberman@microsoft.com>