spl: Add CONFIG_SPL_FIT_SIGNATURE_STRICT
SPL FIT load checks the signature on loadable images, but just continues
in the case of a failure. This is undesirable behavior because the boot
process depends on the authenticity of every loadable.

Adding CONFIG_SPL_FIT_SIGNATURE_STRICT to halt the platform when any
image fails its signature check, including the loadables.

SPL already supports image signature verification, but had no mechanism
to check that the FIT's configuration block was signed correctly.

Add a check near the start of spl_load_simple_fit that verifies the
FIT's configuraiton block, and fails if it's not present or the
signature doesn't match what's stored in the SPL DTB.

Signed-off-by: Henry Beberman <henry.beberman@microsoft.com>