spl: Add Kconfig to block inclusion of SHA1 in SPL
SHA1 is unused in the Windows boot flow, but it's pulled in
automatically with FIT and SPL hashing. All the hashes used are SHA256
and the SHA1 functions account for about 5KB of dead code in the SPL
binary.

Add the option CONFIG_FORBID_SHA1 to stop the SHA1 functions from being
compiled into the binary

Signed-off-by: Henry Beberman <hebeberm@microsoft.com>