grapeboard: Add secret device identity capture
read_and_hide_cyres_identity captures and hides the secret device
identity from the CAAM while the priblob is set to 00. When it's done it
rolls the priblob forward to 11 to prevent anyone else from reading the
value for the rest of boot.

When run during the non-secure boot state it will return the same test
identity as every other system.

Signed-off-by: Henry Beberman <hebeberm@microsoft.com>